+++ src/classes/ControlButton.lua
@@ -38,7 +38,7 @@ local function keybindString(modifierKeys, primaryKey)
                 bindingString = bindingString .. keys:indexToString(primaryKey)
                 deviceString = "Keyboard"
         else
-                bindingString = "Click to Assign"
+                bindingString = ""
         end
 
         return deviceString, bindingString
@@ -52,6 +52,8 @@ function ControlButton:initialize(bind, defaults)
         self.listenerModificators = {}
         self.listenerController = -1
         self.listenerControllerModificator = -1
+        self.listenerInputMode = -1
+        self.inputModeBound = { false, false, false }
 end
 
 function ControlButton:disabled() return self._button:disabled() end
@@ -63,24 +65,33 @@ function ControlButton:boundToController()
         local button = configs.CONTROLS.ini:get(self.bind, "BUTTON", -1)
         local buttonMod = configs.CONTROLS.ini:get(self.bind, "BUTTON_MODIFICATOR", -1)
 
-        if con >= 0 then
+        if con >= 0 and button ~= "" and tonumber(button) ~= -1 then
                 local controllerString = "Joy %s" % con
 
                 if controllers[con] then controllerString = controllers[con].CON end
 
-                return controllerString, buttonString(button + 1, buttonMod + 1)
+                self.inputModeBound[1] = true
+
+                if buttonMod == 0 then return controllerString, string.format("Button %s", button + 1) end
+
+                return controllerString, string.format("Button %s + Button %s", buttonMod + 1, button + 1)
         else
-                return "", "Click to Assign"
+                self.inputModeBound[1] = false
+                return "", ""
         end
 end
 
 function ControlButton:boundToGamepad()
         local button = configs.CONTROLS.ini:get(self.bind, "XBOXBUTTON", "")
 
-        if button and button ~= "-1" and button ~= "" then
-                return "Gamepad", gamepadString(gamepad.codeStringList[button])
+        if button and tonumber(button) ~= -1 and button ~= "" then
+                self.inputModeBound[2] = true
+
+                return "Gamepad", gamepad.codeStringList[button]
         else
-                return "", "Click to Assign"
+                self.inputModeBound[2] = false
+
+                return "", ""
         end
 end
 
@@ -90,7 +101,24 @@ function ControlButton:boundToKey()
 
         if type(modifierKeys[1]) == "table" then modifierKeys = modifierKeys[1] end
 
-        return keybindString(modifierKeys, primaryKey)
+        local deviceString, bindingString = "", ""
+
+        for _, modKey in ipairs(modifierKeys) do
+                if modKey ~= nil and modKey ~= -1 and modKey ~= "-1" and modKey ~= "" then
+                        bindingString = bindingString .. keys.indexStringList[tonumber(modKey)] .. "+"
+                end
+        end
+
+        if primaryKey ~= nil and primaryKey ~= -1 and primaryKey ~= "-1" and primaryKey ~= "" then
+                bindingString = bindingString .. keys:indexToString(primaryKey)
+                deviceString = "Keyboard"
+                self.inputModeBound[3] = true
+        else
+                bindingString = ""
+                self.inputModeBound[3] = false
+        end
+
+        return deviceString, bindingString
 end
 
 function ControlButton:boundTo(inputMode)
@@ -124,6 +152,8 @@ function ControlButton:listenControllerInputs()
                                         self.listenerController = con
                                 end
 
+                                self.listenerInputMode = 1
+
                                 anyButtonDown = true
                         end
                 end
@@ -141,6 +171,8 @@ function ControlButton:listenGamepadInputs()
                                 self.listenerButton = gamepad.indexCodeList[ac.GamepadButton[buttonString]]
 
                                 anyButtonDown = true
+
+                                self.listenerInputMode = 2
                         end
                 end
         end
@@ -172,6 +204,8 @@ function ControlButton:listenKeyboardInputs()
 
                         self.listenerButton = v
 
+                        self.listenerInputMode = 3
+
                         anyButtonDown = true
                 end
         end
@@ -181,18 +215,15 @@ function ControlButton:listenKeyboardInputs()
         return (self.listenerButton ~= -1)
 end
 
-ControlButton.listeners = {
-        ControlButton.listenControllerInputs,
-        ControlButton.listenGamepadInputs,
-        ControlButton.listenKeyboardInputs,
-}
-
 function ControlButton:assignBind(inputMode)
-        local released = self.listeners[inputMode](self)
+        local released = { self:listenControllerInputs(), self:listenGamepadInputs(), self:listenKeyboardInputs() }
 
-        if not released then return false, self.listenerButton, self.listenerModificators end
+        if self.listenerInputMode == -1 or not released[self.listenerInputMode] then
+                return false, self.listenerButton, self.listenerModificators
+        end
 
         return (#self.listenerModificators > 0 or self.listenerButton ~= -1),
+                self.listenerInputMode,
                 self.listenerController,
                 self.listenerButton,
                 self.listenerControllerModificator,
@@ -200,7 +231,7 @@ function ControlButton:assignBind(inputMode)
 end
 
 function ControlButton:clearAssign()
-        self.listenerButton, self.listenerModificators, self.listenerController = -1, {}, -1
+        self.listenerInputMode, self.listenerButton, self.listenerModificators, self.listenerController = -1, -1, {}, -1
 end
 
 function ControlButton:saveBind(inputMode)